chore: rename component file name
