+++ src/pages/dashboard/base/index.vue
@@ -7,7 +7,7 @@
     <!-- 列表排名 -->
     <rank-list class="row-container" />
     <!-- 出入库概览 -->
-    <over-view class="row-container" />
+    <output-overview class="row-container" />
   </div>
 </template>
 
@@ -18,10 +18,10 @@ export default {
 </script>
 
 <script setup lang="ts">
-import TopPanel from './components/top-panel.vue';
-import MiddleChart from './components/middle-chart.vue';
-import RankList from './components/rank-list.vue';
-import OverView from './components/overview.vue';
+import TopPanel from './components/TopPanel.vue';
+import MiddleChart from './components/MiddleChart.vue';
+import RankList from './components/RankList.vue';
+import OutputOverview from './components/OutputOverview.vue';
 </script>
 
 <style scoped>
